moved exception handling to separate file and catching KeyError
